+++ core/src/main/kotlin/org/futurerobotics/temporaryname/control/BangBangController.kt
@@ -1,23 +1,19 @@
 package org.futurerobotics.temporaryname.control
 
 /**
- * A controller that outputs [lessThanOutput] as a signal when the state is less than the reference, [greaterThanOutput] when
- * greater than the reference, and [zeroOutput] if the reference equals the output, and that's it.
+ * A controller that outputs [lessThanOutput] as a signal when the state is less than the reference, [greaterThanOutput]
+ * when greater than the reference, and [zeroOutput] if the reference equals the output, and that's it.
  *
- * One of the most basic controllers one can think of.
+ * One of the more basic controllers.
  */
-/*
-TODO
-Perhaps just ditch this, nonessential
- */
-open class BangBangController<Signal : Any>(
+ class BangBangController<State: Comparable<State>, Signal : Any>(
     private val lessThanOutput: Signal,
     private val greaterThanOutput: Signal,
     private val zeroOutput: Signal
-) : Controller<Double, Double, Signal> {
+) : Controller<State, State, Signal> {
 
     override lateinit var signal: Signal
-    override fun update(reference: Double, currentState: Double, elapsedSeconds: Double) {
+    override fun update(reference: State, currentState: State, elapsedSeconds: Double) {
         val comp = currentState.compareTo(reference)
         signal = when {
             comp < 0 -> lessThanOutput