+++ core/src/main/kotlin/org/futurerobotics/temporaryname/control/PoseEstimation.kt
@@ -4,18 +4,18 @@ import org.futurerobotics.temporaryname.math.Pose2d
 import org.futurerobotics.temporaryname.math.Vector2d
 import org.futurerobotics.temporaryname.math.cosc
 import org.futurerobotics.temporaryname.math.sinc
-import org.futurerobotics.temporaryname.mechanics.PoseMotion
+import org.futurerobotics.temporaryname.mechanics.Motion
 
 /**
  * Non-linearly tracks the global pose, given relative pose velocities.
  */
-class GlobalPoseTracker(initialPose: Pose2d = Pose2d.ZERO) : BaseObserver<PoseMotion, Any, Pose2d>() {
+class GlobalPoseTracker(initialPose: Pose2d = Pose2d.ZERO) : BaseObserver<Motion<Pose2d>, Any, Pose2d>() {
 
     init {
         state = initialPose
     }
 
-    override fun getState(measurement: PoseMotion, lastSignal: Any, elapsedSeconds: Double): Pose2d {
+    override fun getState(measurement: Motion<Pose2d>, lastSignal: Any, elapsedSeconds: Double): Pose2d {
         val (v, dTheta) = measurement.v * elapsedSeconds
         val (x, y) = v
         val sinc = sinc(dTheta)