+++ core/src/main/kotlin/org/futurerobotics/temporaryname/mechanics/MotionStates.kt
@@ -4,7 +4,7 @@ package org.futurerobotics.temporaryname.mechanics
  * Represents the motion of some quantity representing type [T] (e.g. 1d motion, vector2d, pose, linear algebra vector).
  *
  * This contains velocity ([v]), and acceleration ([a])
- * @see [MotionState]
+ * @see [State]
  */
 interface Motion<T : Any> {
 
@@ -26,41 +26,40 @@ operator fun <T : Any> Motion<T>.component2(): T = a
  * This is a subclass of:
  * @see [Motion]
  */
-interface MotionState<T : Any> : Motion<T> {
+interface State<T : Any> : Motion<T> {
 
-    /** The position of this [MotionState] */
+    /** The position of this [State] */
     val s: T
-    /** The velocity of this [MotionState] */
+    /** The velocity of this [State] */
     override val v: T
-    /** The acceleration of this [MotionState] */
+    /** The acceleration of this [State] */
     override val a: T
 }
 
-operator fun <T : Any> MotionState<T>.component1(): T = s
-operator fun <T : Any> MotionState<T>.component2(): T = v
-operator fun <T : Any> MotionState<T>.component3(): T = a
+operator fun <T : Any> State<T>.component1(): T = s
+operator fun <T : Any> State<T>.component2(): T = v
+operator fun <T : Any> State<T>.component3(): T = a
 
 /** An simple implementation of [Motion] that holds values in fields */
 open class ValueMotion<T : Any>(override val v: T, override val a: T) : Motion<T> {
 
     override fun equals(other: Any?): Boolean = when {
         this === other -> true
-        other !is ValueMotionState<*> -> false
+        other !is ValueState<*> -> false
         else -> v == other.v && a == other.a
     }
 
     override fun hashCode(): Int = 31 * v.hashCode() + a.hashCode()
 }
 
-/** An simple implementation of [MotionState] that holds values in fields */
-open class ValueMotionState<T : Any>(override val s: T, v: T, a: T) : ValueMotion<T>(v, a), MotionState<T> {
+/** An simple implementation of [State] that holds values in fields */
+open class ValueState<T : Any>(override val s: T, v: T, a: T) : ValueMotion<T>(v, a), State<T> {
 
     override fun equals(other: Any?): Boolean = when {
         this === other -> true
-        other !is ValueMotionState<*> -> false
+        other !is ValueState<*> -> false
         else -> super.equals(other) && s == other.s
     }
 
     override fun hashCode(): Int = 31 * super.hashCode() + s.hashCode()
 }
-